+++ src/visualization/annotation/image_classification.cpp
@@ -50,9 +50,9 @@ annotate_spec::Data ImageClassification::getItems(size_t start, size_t end) {
     flex_image img = flex_data.at(i).get<flex_image>();
     img = turi::image_util::encode_image(img);
 
-    size_t img_width = img.m_width;
-    size_t img_height = img.m_height;
-    size_t img_channels = img.m_channels;
+    int img_width = static_cast<int>(img.m_width);
+    int img_height = static_cast<int>(img.m_height);
+    int img_channels = static_cast<int>(img.m_channels);
 
     annotate_spec::Datum *datum = data.add_data();
     annotate_spec::ImageDatum *img_datum = datum->add_images();
@@ -197,9 +197,9 @@ annotate_spec::Similarity ImageClassification::get_similar_items(size_t index,
     flex_image img = data_sa[idx];
     img = turi::image_util::encode_image(img);
 
-    size_t img_width = img.m_width;
-    size_t img_height = img.m_height;
-    size_t img_channels = img.m_channels;
+    int img_width = static_cast<int>(img.m_width);
+    int img_height = static_cast<int>(img.m_height);
+    int img_channels = static_cast<int>(img.m_channels);
 
     annotate_spec::Datum *datum = similar.add_data();
     annotate_spec::ImageDatum *img_datum = datum->add_images();
@@ -263,7 +263,7 @@ bool ImageClassification::setAnnotations(
 
     switch (label.labelIdentifier_case()) {
     case annotate_spec::Label::LabelIdentifierCase::kIntLabel:
-      _addAnnotationToSFrame(sf_idx, label.intlabel());
+      _addAnnotationToSFrame(sf_idx, static_cast<int>(label.intlabel()));
       break;
     case annotate_spec::Label::LabelIdentifierCase::kStringLabel:
       _addAnnotationToSFrame(sf_idx, label.stringlabel());
